Rename multi_tasks namespace to multi_task
