+++ autoware_ml/configs/datamodule/t4dataset/datasets/detection3d/default_detection3d.yaml
@@ -1,4 +1,4 @@
-_target_: autoware_ml.datamodule.multi_tasks.t4dataset.detection3d.T4Detection3DTask
+_target_: autoware_ml.datamodule.multi_task.t4dataset.detection3d.T4Detection3DTask
 _convert_: all
 
 # Always set to None since we run splitter in the process
