+++ autoware_ml/configs/datamodule/t4dataset/datasets/transforms/default_train_transforms.yaml
@@ -1,8 +1,8 @@
-_target_: autoware_ml.transforms.multi_tasks.base.MultiTaskTransformsCompose
+_target_: autoware_ml.transforms.multi_task.base.MultiTaskTransformsCompose
 _convert_: all
 
 pipeline:
-  - _target_: autoware_ml.transforms.multi_tasks.point_cloud.loading.LoadPointsFromFile
+  - _target_: autoware_ml.transforms.multi_task.point_cloud.loading.LoadPointsFromFile
     _convert_: all
 
     load_dim: ${database.lidar_pointcloud_num_features}
