+++ autoware_ml/datamodule/multi_task/t4dataset/detection3d.py
@@ -3,9 +3,9 @@ import polars as pl
 
 from autoware_ml.databases.schemas.dataset_schemas import DatasetTableSchema
 from autoware_ml.databases.schemas.box3d_schemas import Box3DDatasetSchema
-from autoware_ml.datamodule.multi_tasks.base_dataset_task import BaseDatasetTask
-from autoware_ml.datamodule.multi_tasks.dataclasses.detection3d import Detection3DGTSample
-from autoware_ml.datamodule.multi_tasks.dataclasses.multi_task_samples import MultiTaskGTSample
+from autoware_ml.datamodule.multi_task.base_dataset_task import BaseDatasetTask
+from autoware_ml.datamodule.multi_task.dataclasses.detection3d import Detection3DGTSample
+from autoware_ml.datamodule.multi_task.dataclasses.multi_task_samples import MultiTaskGTSample
 
 
 class T4Detection3DTask(BaseDatasetTask):
